restructure upvalues, use Rc internally
Move UpValues entirely to safe code and fix a regression that occured
in the migration to broom. By using Rc internally we avoid the
issue which copied the internal value instead of referring to the same
object from the closure, and we also avoid the use of unsafe code at the
cost of some runtime complexity.